bot-scrap/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
mysql-connector-python==8.2.0
python-dotenv==1.0.0
schedule==1.2.0
colorama==0.4.6 